requirements.txt
mistralai
python-dotenv
requests
faiss-cpu